Allow passing fixed ID to create build channel
Previously passing an ID to POST /build_channels was required but the
attribute was ignored and instead a random UUID was used instead.

Since other endpoints allow passing a fixed ID, support this for this
endpoint also to allow idempotent create requests.

At the same time also remove the required attribute from the API, so
that if the field is not present in the request json that still a new
UUID is generated.

Change-Id: I1064792b577fd37221a605bb112d6ceecc10c969